# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meal Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scheduled-items:
    get:
      operationId: listScheduledItems
      tags: [schedule]
      summary: List all scheduled recurring grocery items
      responses:
        "200":
          description: Scheduled items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledItem"

    post:
      operationId: createScheduledItem
      tags: [schedule]
      summary: Create a recurring scheduled grocery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduledItemBody"
      responses:
        "201":
          description: Created scheduled item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledItem"
  /scheduled-items/{id}:
    patch:
      operationId: updateScheduledItem
      tags: [schedule]
      summary: Update a scheduled item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduledItemBody"
      responses:
        "200":
          description: Updated scheduled item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledItem"

    delete:
      operationId: deleteScheduledItem
      tags: [schedule]
      summary: Delete a scheduled item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ScheduledItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        quantity:
          type: string
        unit:
          type: string
          nullable: true
        category:
          type: string
        scheduleType:
          type: string
          description: "weekly | biweekly | every_other_day | custom"
        scheduleDaysInterval:
          type: integer
          nullable: true
        nextDueDate:
          type: string
          format: date
        lastAddedDate:
          type: string
          format: date
          nullable: true
        isActive:
          type: boolean
      required:
        [
          id,
          name,
          quantity,
          category,
          scheduleType,
          nextDueDate,
          isActive,
        ]
    CreateScheduledItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
        unit:
          type: string
          nullable: true
        category:
          type: string
        scheduleType:
          type: string
          description: "weekly | biweekly | every_other_day | custom"
        scheduleDaysInterval:
          type: integer
          nullable: true
          description: Required if scheduleType is 'custom'
      required: [name, quantity, category, scheduleType]
...
    UpdateScheduledItemBody:
      type: object
      properties:
        name:
          type: string
          nullable: true
        quantity:
          type: string
          nullable: true
        scheduleType:
          type: string
          nullable: true
        scheduleDaysInterval:
          type: integer
          nullable: true
        isActive:
          type: boolean
          nullable: true
        nextDueDate:
          type: string
          format: date
          nullable: true

    # ── BILLING ─────────────────────────────────────────────────────────────
...